用户相关

+++ duobao-backend/src/main/java/team/wuxie/crowdfunding/service/CurrentUserService.java
@@ -13,5 +13,5 @@ import team.wuxie.crowdfunding.domain.CurrentUser;
  */
 public interface CurrentUserService {
 
-    boolean canAccessUser(CurrentUser currentUser, Long userId);
+    boolean canAccessUser(CurrentUser currentUser, Integer userId);
 }
